Backend/requirements.txt
Flask==3.1.1
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
email-validator==2.2.0
python-dotenv==1.1.0
pandas==2.2.3
numpy==2.2.6
langchain==0.3.25
langchain-community==0.3.24
langchain-openai==0.3.18
langchain-text-splitters==0.3.8
langchain-chroma==0.2.4
openai==1.82.0
werkzeug==3.1.3
